Contodminio id Integer

+++ src/main/java/com/br/Kodominio/dao/ICondominio.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 @EnableJpaRepositories
-public interface ICondominio extends CrudRepository<Condominio, Long> {
+public interface ICondominio extends CrudRepository<Condominio, Integer> {
 
 }
